working api_routes

+++ routes/apiRoutes.js
@@ -1,6 +1,6 @@
 var db = require("../models");
 
-module.exports = function(app) {
+module.exports = function(sequelize, app) {
 
   // register
   app.post("/api/register", function(req, res) {
@@ -16,7 +16,8 @@ module.exports = function(app) {
   });
 
   // login
-  app.get("/api/login", function(req, res) {
+  app.post("/api/login", function(req, res) {
+    console.log(req.body);
     db.User.findOne({
       where: {
         username: req.body.username,
@@ -46,10 +47,10 @@ module.exports = function(app) {
   app.post("/api/buy", function(req, res) {
     console.log(req.body);
     db.Portfolio.create({
-      userid:   req.body.userid,
-      ticker:   req.body.ticker,
-      quantity: req.body.quantity,
-      day:      req.body.day
+      UserId:         req.body.userid,
+      CompanyId:      req.body.companyid,
+      stock_quantity: req.body.stock_quantity,
+      buy_price:      req.body.buy_price
     }).then(function(dbPortfolio) {
         res.json(dbPortfolio);
     }).catch(function(err) {
@@ -57,15 +58,46 @@ module.exports = function(app) {
     });
   });
 
+  //Model.update({ field: sequelize.literal('field + 2') }, { where: { id: model_id } });
+
   // sell 1 stock (put portfolio table)
-  app.delete("/api/sell", function(req, res) {
-    db.Portfolio.destroy({
+  app.put("/api/sell", function(req, res) {
+    var sell_qty =  req.body.stock_quantity;
+    var literal = "stock_quantity - " + sell_qty; // "stock_quantity - 30"
+    console.log(literal);
+    db.Portfolio.update({
+      stock_quantity: sequelize.literal(literal)
+    }, {
       where: {
-        ticker: req.body.ticker
+        id: req.body.id
       }
-    }).then(function(db) {
-      res.json(dbPortfolio)
+    }).then(function(dbPortfolio) {
+      // Get new stock quantity, since the update only return number of rows updated
+      console.log("************" + JSON.stringify(dbPortfolio));
+      db.Portfolio.findOne({
+        where: {
+          id: req.body.id
+        }
+      }).then(function(dbGetRes) {
+        // If the user sells all stocks for a particular company (quantity = 0), 
+        // then delete portfolio row for that company
+        if (dbGetRes.stock_quantity <= 0) {
+          db.Portfolio.destroy({
+            where: {
+              id: dbGetRes.id
+            }
+          }).then(function(dbDelRes) {
+            res.json(dbGetRes)
+          }).catch(function(err) {
+            res.json(err);
+          });
+        } else {
+          res.json(dbGetRes);
+        }
+      });
     }).catch(function(err) {
-      res.json(err);
-    });
+        res.json(err);
+      });
   });
+}
+
